@@ -1,15 +1,15 @@
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:injectable/injectable.dart';
 
+@lazySingleton
 class AuthHelper {
-  static final instance = AuthHelper(FirebaseAuth.instance);
-  final FirebaseAuth _auth;
 
-  AuthHelper(this._auth);
+  AuthHelper();
 
   Future<UserCredential> loginWithEmailAndPassword(
       String emailAddress, String password) async {
     try {
-      return await _auth.signInWithEmailAndPassword(
+      return await FirebaseAuth.instance.signInWithEmailAndPassword(
         email: emailAddress,
         password: password,
       );
@@ -27,7 +27,7 @@ class AuthHelper {
     String password,
   ) async {
     try {
-      return await _auth.createUserWithEmailAndPassword(
+      return await FirebaseAuth.instance.createUserWithEmailAndPassword(
         email: emailAddress,
         password: password,
       );
@@ -41,20 +41,20 @@ class AuthHelper {
   }
 
   Stream<bool> isSignedInStream() {
-    return _auth.authStateChanges().map((event) => event != null);
+    return FirebaseAuth.instance.authStateChanges().map((event) => event != null);
   }
 
   bool isSignedIn() {
-    return _auth.currentUser != null;
+    return FirebaseAuth.instance.currentUser != null;
   }
 
   String? getUserId() {
-    return _auth.currentUser?.uid;
+    return FirebaseAuth.instance.currentUser?.uid;
   }
 
   Future<void> logout() async {
     try {
-      return await _auth.signOut();
+      return await FirebaseAuth.instance.signOut();
     } catch (e) {
       throw UnableToLogout();
     }
