@@ -1,4 +1,3 @@
-import 'package:animal_rescue/di/get_it.dart';
 import 'package:animal_rescue/gen/colors.gen.dart';
 import 'package:animal_rescue/utils/global_observer.dart';
 import 'package:firebase_core/firebase_core.dart';
@@ -7,13 +6,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 
 import 'arch/presentation/core/app.dart';
+import 'di/di.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
-  setupDi();
+  configureDependencies();
   Bloc.observer = GlobalObserver();
-  runApp(App());
+  runApp(const App());
   configLoading();
 }
 
